Ensure that we clean the supervisor.d dir
... so that we can effectively unmanage processes that were under
supervisor control.

This utilises the salt 'clean: True' technique, where we make the clean
directory require (via require_in) the files that we want to remain in the
cleaned dir. This is logically weird, since naturally these files depend on this
directory. Adding 'makedirs: True' to each of the required files works around this
paradox.